requirements.txt
pandas
matplotlib
seaborn
plotly
squarify
sklearn
jupyter
gensim
vaderSentimentx
nltk
lightGBM